[Concurrency] Move concurrency-specific code into its own file.
